+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/factories/GaaSRestExceptionFactoryTest.java
@@ -19,13 +19,13 @@ package uk.gov.gchq.gaffer.gaas.factories;
 import io.kubernetes.client.openapi.ApiException;
 import org.junit.jupiter.api.Test;
 import uk.gov.gchq.gaffer.gaas.exception.GaaSRestApiException;
-import uk.gov.gchq.gaffer.gaas.utilities.UnitTest;
+import uk.gov.gchq.gaffer.gaas.util.UnitTest;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static uk.gov.gchq.gaffer.gaas.utilities.ApiExceptionTestFactory.makeApiException_custom;
-import static uk.gov.gchq.gaffer.gaas.utilities.ApiExceptionTestFactory.makeApiException_duplicateGraph;
-import static uk.gov.gchq.gaffer.gaas.utilities.ApiExceptionTestFactory.makeApiException_loggedOutOfCluster;
-import static uk.gov.gchq.gaffer.gaas.utilities.ApiExceptionTestFactory.makeApiException_timeout;
+import static uk.gov.gchq.gaffer.gaas.util.ApiExceptionTestFactory.makeApiException_custom;
+import static uk.gov.gchq.gaffer.gaas.util.ApiExceptionTestFactory.makeApiException_duplicateGraph;
+import static uk.gov.gchq.gaffer.gaas.util.ApiExceptionTestFactory.makeApiException_loggedOutOfCluster;
+import static uk.gov.gchq.gaffer.gaas.util.ApiExceptionTestFactory.makeApiException_timeout;
 
 @UnitTest
 public class GaaSRestExceptionFactoryTest {
