+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/util/ApiExceptionTestFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.gchq.gaffer.gaas.utilities;
+package uk.gov.gchq.gaffer.gaas.util;
 
 import io.kubernetes.client.openapi.ApiException;
 import java.util.Arrays;
